fixed typo

+++ core/src/com/fixthewall/game/Game.java
@@ -25,7 +25,7 @@ public class Game extends com.badlogic.gdx.Game {
     public FitViewport viewport;
     public AssetManager ass;
 
-    public static final float GAME_WIDTH = 1080;
+    public static final int GAME_WIDTH = 1080;
     public static final int GAME_HEIGHT = 1920;
 
 	public MusicLogic playlist;
